[fix] 예매내역 Branch 분리
