+++ src/routes/orderlistRouter.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const orderlistController = require('../controllers/orderlistController');
+const router = express.Router();
+
+router.get("/check-order/:userId", orderlistController.CheckOrderController)
+
+
+module.exports = router;
